requirements.txt
streamlit
streamlit-option-menu
firebase-admin
pandas
plotly
openai
httpx
fpdf2
openpyxl
